refactoring: Change logger to protocol
